Put all classes into hermione package
