--- src/main/java/exceptions/DateUtilsException.java
@@ -1,7 +1,0 @@
-package exceptions;
-
-public class DateUtilsException extends RuntimeException {
-    public DateUtilsException(String message) {
-        super("[ERROR] Date Error: " + message);
-    }
-}
