--- src/main/java/exceptions/InvalidCommandException.java
@@ -1,7 +1,0 @@
-package exceptions;
-
-public class InvalidCommandException extends RuntimeException {
-    public InvalidCommandException(String message) {
-        super("[ERROR] Invalid command: " + message);
-    }
-}
