--- src/main/java/exceptions/TaskValidationException.java
@@ -1,7 +1,0 @@
-package exceptions;
-
-public class TaskValidationException extends RuntimeException {
-    public TaskValidationException(String message) {
-        super("[ERROR] Task Error: " + message);
-    }
-}
