+++ src/main/java/hermione/commands/Command.java
@@ -1,0 +1,16 @@
+package hermione.commands;
+
+import hermione.storage.TaskStorage;
+
+public abstract class Command {
+
+    protected final TaskStorage storage;
+    protected final String argument;
+
+    public Command(TaskStorage storage, String argument) {
+        this.storage = storage;
+        this.argument = argument;
+    }
+
+    public abstract String execute();
+}
