+++ src/main/java/hermione/exceptions/DateUtilsException.java
@@ -1,0 +1,7 @@
+package hermione.exceptions;
+
+public class DateUtilsException extends RuntimeException {
+    public DateUtilsException(String message) {
+        super("[ERROR] Date Error: " + message);
+    }
+}
