+++ src/main/java/hermione/exceptions/InvalidCommandException.java
@@ -1,0 +1,7 @@
+package hermione.exceptions;
+
+public class InvalidCommandException extends RuntimeException {
+    public InvalidCommandException(String message) {
+        super("[ERROR] Invalid command: " + message);
+    }
+}
