+++ src/main/java/hermione/exceptions/TaskValidationException.java
@@ -1,0 +1,7 @@
+package hermione.exceptions;
+
+public class TaskValidationException extends RuntimeException {
+    public TaskValidationException(String message) {
+        super("[ERROR] Task Error: " + message);
+    }
+}
