+++ src/main/java/hermione/tasks/TaskList.java
@@ -1,0 +1,41 @@
+package hermione.tasks;
+
+import java.util.List;
+import java.util.stream.Collectors;
+import java.util.stream.IntStream;
+
+public class TaskList {
+
+    private List<Task> tasks;
+
+    public TaskList(List<Task> tasks) {
+        this.tasks = tasks;
+    }
+
+    public void addTask(Task task) {
+        this.tasks.add(task);
+    }
+
+    public void deleteTask(Task task) {
+        this.tasks.remove(task);
+    }
+
+    public Task getTask(int index) {
+        return this.tasks.get(index);
+    }
+
+    public List<Task> getTasks() {
+        return this.tasks;
+    }
+
+    public int getSize() {
+        return this.tasks.size();
+    }
+
+    @Override
+    public String toString() {
+        return IntStream.range(0, tasks.size())
+                .mapToObj(i -> (i + 1) + ". " + tasks.get(i).toString())
+                .collect(Collectors.joining("\n"));
+    }
+}
