--- src/main/java/storage/TaskStorage.java
@@ -1,11 +1,0 @@
-package storage;
-
-import tasks.Task;
-import tasks.TaskList;
-
-public interface TaskStorage {
-    TaskList getTasks();
-    void addTask(Task task);
-    void setTaskCompletion(Task task, boolean isComplete);
-    Task deleteTask(int index);
-}
