--- src/main/java/tasks/ToDo.java
@@ -1,12 +1,0 @@
-package tasks;
-
-public class ToDo extends Task{
-    public ToDo(String description, boolean isCompleted) {
-        super(description, isCompleted);
-    }
-
-    @Override
-    public String toString() {
-        return "[T]" + super.toString();
-    }
-}
