+++ text-ui-test/runtest.sh
@@ -13,13 +13,13 @@ then
 fi
 
 # compile the code into the bin folder, terminates if error occurred
-if ! javac -cp ../src/main/java -Xlint:none -d ../bin ../src/main/java/*.java
+if ! javac -cp ../src/main/java -Xlint:none -d ../bin ../src/main/java/hermione/*.java
 then
     echo "********** BUILD FAILURE **********"
     exit 1
 fi
 
-# run the program, feed commands from input.txt file and redirect the output to the ACTUAL.TXT
+# run the program, feed hermione.commands from input.txt file and redirect the output to the ACTUAL.TXT
 java -classpath ../bin Hermione < input.txt > ACTUAL.TXT
 
 # convert to UNIX format
